Fix boolean inversion in VBE screen clear.
From the VESA VBE doc: "If bit D15 is not set, all reported image pages,
based on Function 00h returned information NumberOfImagePages, will be
cleared to 00h in graphics mode, and 20 07 in text mode. Memory over and
above the reported image pages will not be changed. If bit D15 is set,
then the contents of the frame buffer after the mode change is undefined."

D15 here means the high bit of %bx.