config: Move compiler flags from configure.ac to Makefile.am
Remove unused XORG_INCS variable.
Remove unused INCLUDES='-I$(top_srcdir)/src'
This statement is redundant and not used in the makefile

Remove unrequired '-I$(top_srcdir)/src'
The current dir is already included by default in the makefile
top_builddir = ..
DEFAULT_INCLUDES = -I. -I$(top_builddir)

Move compiler flags from configure.ac to Makefile.am
CFLAGS is an automake defined variable that should not be set
by the module. It should not be AC_SUBST either, it already is.
Use AM_CFLAGS in Makefile.am. This will allow the user to override
the flags as they will be in the right order.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>